feat(js): extract composable web and browser tools
