+++ examples/cloudflare-workers/README.md
@@ -155,6 +155,13 @@ forbid `eval` and `new Function`. This retains Nanocodex's typed Rust tool
 lifecycle and caller-defined handlers without shipping a JavaScript evaluator.
 Node-based consumers may continue to use Code Mode when their host permits it.
 
+Set `WEB_TOOL_URL` to install the standard `web__run` definition from
+`nanocodex/tools` directly in each Durable Object agent. The binding repairs
+common malformed model argument shapes and posts `{ commands, session_id }` to
+that HTTPS endpoint; the Worker retains endpoint credentials in the optional
+`WEB_TOOL_TOKEN` secret. No browser OPFS or shell code enters this server-side
+Worker graph.
+
 ## Validate and deploy
 
 ```sh
