+++ js/bindings/tools/browser/devTunnelsSshBrowser.mjs
@@ -1,23 +1,13 @@
 // Browser-only barrel for dev-tunnels-ssh. The upstream CommonJS root eagerly
 // evaluates Node stream classes even though its crypto implementation correctly
 // selects Web Crypto. Export only the browser-safe protocol surface we use.
-export {
-  SshAlgorithms,
-  Encryption,
-  ECDsa,
-  Rsa,
-} from "@microsoft/dev-tunnels-ssh/algorithms/sshAlgorithms.js";
+export { SshAlgorithms, Encryption, ECDsa, Rsa, } from "@microsoft/dev-tunnels-ssh/algorithms/sshAlgorithms.js";
 export { SshDataReader } from "@microsoft/dev-tunnels-ssh/io/sshData.js";
 export { DerReader, DerWriter } from "@microsoft/dev-tunnels-ssh/io/derData.js";
 export { BigInt } from "@microsoft/dev-tunnels-ssh/io/bigInt.js";
-
-export {
-  CancellationToken,
-  CancellationTokenSource,
-} from "@microsoft/dev-tunnels-ssh/util/cancellation.js";
+export { CancellationToken, CancellationTokenSource, } from "@microsoft/dev-tunnels-ssh/util/cancellation.js";
 export { CommandRequestMessage } from "@microsoft/dev-tunnels-ssh/messages/connectionMessages.js";
 export { SshAuthenticationType } from "@microsoft/dev-tunnels-ssh/events/sshAuthenticatingEventArgs.js";
 export { SshClientSession } from "@microsoft/dev-tunnels-ssh/sshClientSession.js";
 export { SshSessionConfiguration } from "@microsoft/dev-tunnels-ssh/sshSessionConfiguration.js";
 export { WebSocketStream } from "@microsoft/dev-tunnels-ssh/streams.js";
-export type { SshChannelClosedEventArgs } from "@microsoft/dev-tunnels-ssh/events/sshChannelClosedEventArgs.js";
