+++ js/bindings/tools/browser/unsupportedNodeRsa.mjs
@@ -1,0 +1,5 @@
+export default class UnsupportedNodeRsa {
+    constructor() {
+        throw new Error("node-rsa is unavailable in browsers; Web Crypto must be used");
+    }
+}
