+++ js/bindings/tools/index.mjs
@@ -1,0 +1,6 @@
+export {
+  imageGeneration,
+  updatePlan,
+  viewImage,
+  web,
+} from "./standard.mjs";
