+++ js/bindings/tools/vite.d.mts
@@ -1,0 +1,8 @@
+export type NanocodexToolsVitePlugin = Readonly<{
+  name: "nanocodex-tools";
+  enforce: "pre";
+  resolveId(source: string): string | null;
+}>;
+
+/** Keeps unreachable Node-only SSH fallbacks out of browser and Worker bundles. */
+export function nanocodexTools(): NanocodexToolsVitePlugin;
