+++ web/package.json
@@ -23,9 +23,6 @@
   },
   "dependencies": {
     "@cloudflare/containers": "0.3.7",
-    "@eduoj/wasm-clang": "1.0.0",
-    "@microsoft/dev-tunnels-ssh": "3.12.40",
-    "@microsoft/dev-tunnels-ssh-keys": "3.12.40",
     "@modelcontextprotocol/sdk": "1.30.0",
     "@noble/hashes": "1.8.0",
     "@pierre/diffs": "1.2.12",
@@ -36,11 +33,9 @@
     "@tanstack/react-query": "^5.101.4",
     "@tanstack/react-virtual": "^3.14.6",
     "accounts": "^0.16.0",
-    "buffer": "6.0.3",
     "diff": "^9.0.0",
     "htm": "3.1.1",
     "isomorphic-git": "^1.41.5",
-    "just-bash": "^3.4.0",
     "lucide-react": "^1.24.0",
     "minisearch": "7.2.0",
     "mppx": "^0.8.15",
@@ -50,7 +45,6 @@
     "nanocodex-tui": "file:../js/tui",
     "nanocodex-tui-react": "file:../js/tui-react",
     "ox": "^0.14.33",
-    "pyodide": "314.0.5",
     "react": "19.2.6",
     "react-dom": "19.2.6",
     "react-router": "^7.18.2",
